Last update to project progress report...
Updated README file (attributing Craig Reynolds Flocking code)

Corrected some typos and added comments

This is the last update made in this "project-progress-report" folder. The future updates will be made in the "final" folder of project 2 (so go check out the projects folder for updates on this project)

Why I doing this? Because I literally forgot I was working on this project in the wrong folder so transferring it now before it becomes more confusing